requirements.txt
requests==2.31.0
google-cloud-pubsub==2.20.0
google-cloud-bigquery==3.12.0
google-cloud-storage==2.14.0
torch
python-dotenv==1.0.0
together==0.1.0
google-cloud-aiplatform==1.35.0
google-cloud-secret-manager==2.16.0
google-cloud-translate==3.16.0
